requirements.txt
asgiref==3.8.1
attrs==23.2.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
distro==1.9.0
Django==4.2.11
djangorestframework==3.15.1
docker==7.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
gunicorn==22.0.0
idna==3.7
jsonschema==3.2.0
packaging==24.0
paramiko==3.4.0
pycparser==2.22
PyNaCl==1.5.0
pyrsistent==0.20.0
python-dotenv==0.21.1
PyYAML==5.4.1
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
texttable==1.7.0
typing_extensions==4.11.0
urllib3==2.2.1
websocket-client==0.59.0
